// src/types/pronosticiParams.types.ts
import { Avvenimento, Disciplina, Manifestazione } from "src/components/prematch/prematch-api";

export type PronosticiParamsType = {
  codicePalinsesto: number;
  codiceAvvenimento: number;
  codiceScommessa: number;
  codiceEsito: number;
  codiceEsitoAAMS?: number;
  codiceClasseEsito: number;
  codiceClasseEsitoAAMS?: number;
  codiceDisciplina: number;
  quota: number;
  descrizioneAvvenimento: string;
  codiceManifestazione: number;
  descrizioneScommessa: string;
  descrizioneEsito: string;
  idInfoAggiuntiva: number;
  legameMultipla: number;
  legameMinimo: number;
  legameMassimo: number;
  multipla: number;
  blackListMin: number;
  blackListMax: number;
  live: boolean;
  isFissa?: boolean;
  formattedDataAvvenimento: string;
  dataAvvenimento: string;
  siglaDisciplina: string;
  siglaManifestazione: string;
  avvenimento: Avvenimento | undefined;
  manifestazione: Manifestazione | undefined;
  disciplina: Disciplina | undefined;
  tieResult?: number;
  tieOdds?: number;
  tieOddsFormatted?: string;
  descrizioneManifestazione?: string;
  handicap?: number;
  handicapVariato?: boolean;
  liveDelay?: number;
  providerId?: number;
};

export type PronosticiParamsVirtualType = {
  eventId: number;
  dataAvvenimento: string;
  formattedDataAvvenimento: string;
  descrizioneEvento: string;
  formattedOrario: string;
  codiceDisciplina: number;
  siglaDisciplina: string;
  codicePalinsesto: string;
  codiceAvvenimento: string;
  provider: number;
  rtp: number;
  result: null;
  descrizioneAvvenimento: string;
  descrizioneEsito: string;
  quota: number;
  formattedQuota: string;
  probwin: number;
  codiceScommessa: number;
  stato: number;
  codiceEsito: number;
};

export type PronosticiResponseType = {
  codicePalinsesto: number;
  codiceAvvenimento: number;
  codiceScommessa: number;
  codiceEsito: number;
  codiceClasseEsito: number;
  codiceDisciplina: number;
  quota: number;
  descrizioneAvvenimento: string;
  siglaDisciplina: string;
  codiceManifestazione: number;
  siglaManifestazione: string;
  blackListMax?: number;
  blackListMin?: number;
  codiceClasseEsitoAAMS: number;
  codiceEsitoAAMS: number;
  dataAvvenimento: string;
  descrizioneEsito: string;
  descrizioneManifestazione: string;
  descrizioneScommessa: string;
  fissa: boolean;
  formattedDataAvvenimento: string;
  handicap: number;
  handicapVariato: boolean;
  id: number;
  legameMassimo: number;
  legameMinimo: number;
  legameMultipla: number;
  live: boolean;
  multipla: number;
  providerId: number;
  quotaVariata: boolean;
  stato: number;
  tieResult: number;
  idInfoAggiuntiva: number;
};